rename ottltraces to ottlspan
