+++ processor/routingprocessor/traces.go
@@ -24,7 +24,7 @@ import (
 	"go.uber.org/multierr"
 	"go.uber.org/zap"
 
-	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/ottl/contexts/ottltraces"
+	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/ottl/contexts/ottlspan"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/processor/routingprocessor/internal/common"
 )
 
@@ -35,7 +35,7 @@ type tracesProcessor struct {
 	config *Config
 
 	extractor extractor
-	router    router[component.TracesExporter, ottltraces.TransformContext]
+	router    router[component.TracesExporter, ottlspan.TransformContext]
 }
 
 func newTracesProcessor(settings component.TelemetrySettings, config component.ProcessorConfig) *tracesProcessor {
@@ -44,11 +44,11 @@ func newTracesProcessor(settings component.TelemetrySettings, config component.P
 	return &tracesProcessor{
 		logger: settings.Logger,
 		config: cfg,
-		router: newRouter[component.TracesExporter, ottltraces.TransformContext](
+		router: newRouter[component.TracesExporter, ottlspan.TransformContext](
 			cfg.Table,
 			cfg.DefaultExporters,
 			settings,
-			ottltraces.NewParser(common.Functions[ottltraces.TransformContext](), settings),
+			ottlspan.NewParser(common.Functions[ottlspan.TransformContext](), settings),
 		),
 		extractor: newExtractor(cfg.FromAttribute, settings.Logger),
 	}
@@ -92,7 +92,7 @@ func (p *tracesProcessor) route(ctx context.Context, t ptrace.Traces) error {
 	var errs error
 	for i := 0; i < t.ResourceSpans().Len(); i++ {
 		rspans := t.ResourceSpans().At(i)
-		stx := ottltraces.NewTransformContext(
+		stx := ottlspan.NewTransformContext(
 			ptrace.Span{},
 			pcommon.InstrumentationScope{},
 			rspans.Resource(),
