+++ processor/transformprocessor/internal/common/traces.go
@@ -24,13 +24,13 @@ import (
 	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/ottl"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/ottl/contexts/ottlresource"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/ottl/contexts/ottlscope"
+	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/ottl/contexts/ottlspan"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/ottl/contexts/ottlspanevent"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/ottl/contexts/ottltraces"
 )
 
 var _ consumer.Traces = &traceStatements{}
 
-type traceStatements []*ottl.Statement[ottltraces.TransformContext]
+type traceStatements []*ottl.Statement[ottlspan.TransformContext]
 
 func (t traceStatements) Capabilities() consumer.Capabilities {
 	return consumer.Capabilities{
@@ -45,7 +45,7 @@ func (t traceStatements) ConsumeTraces(ctx context.Context, td ptrace.Traces) er
 			sspans := rspans.ScopeSpans().At(j)
 			spans := sspans.Spans()
 			for k := 0; k < spans.Len(); k++ {
-				tCtx := ottltraces.NewTransformContext(spans.At(k), sspans.Scope(), rspans.Resource())
+				tCtx := ottlspan.NewTransformContext(spans.At(k), sspans.Scope(), rspans.Resource())
 				for _, statement := range t {
 					_, _, err := statement.Execute(ctx, tCtx)
 					if err != nil {
@@ -94,7 +94,7 @@ func (s spanEventStatements) ConsumeTraces(ctx context.Context, td ptrace.Traces
 
 type TraceParserCollection struct {
 	parserCollection
-	traceParser     ottl.Parser[ottltraces.TransformContext]
+	traceParser     ottl.Parser[ottlspan.TransformContext]
 	spanEventParser ottl.Parser[ottlspanevent.TransformContext]
 }
 
@@ -107,7 +107,7 @@ func NewTraceParserCollection(functions map[string]interface{}, settings compone
 			resourceParser: ottlresource.NewParser(ResourceFunctions(), settings),
 			scopeParser:    ottlscope.NewParser(ScopeFunctions(), settings),
 		},
-		traceParser: ottltraces.NewParser(functions, settings),
+		traceParser: ottlspan.NewParser(functions, settings),
 	}
 
 	for _, op := range options {
