+++ processor/transformprocessor/internal/traces/functions_test.go
@@ -20,12 +20,12 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
-	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/ottl/contexts/ottltraces"
+	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/ottl/contexts/ottlspan"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/processor/transformprocessor/internal/common"
 )
 
 func Test_DefaultFunctions(t *testing.T) {
-	expected := common.Functions[ottltraces.TransformContext]()
+	expected := common.Functions[ottlspan.TransformContext]()
 	actual := Functions()
 	require.Equal(t, len(expected), len(actual))
 	for k := range actual {
